Remove styles for magellan's 'active' class
The default behavior of magellan is slightly bugged, in that it defaults
to adding the 'active' class to the first magellan item immediately on
page load. The class then does not update as the user scrolls, causing
the first link to be highlighted even when it is not the active section.

I had styled these links for magellan's 'active' class, but am now
removing those styles as they would have been confusing for users.